CI/TST: Cleanups (#54362)
